backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-core
